git-svn-id: https://mattari.mydns.jp/repos/android_apps/trunk/NitechBBSrv@608 ebda3f91-3b32-42d3-bb0a-b50f2d3b12aa

+++ app/models/entity/User.java
@@ -117,7 +117,7 @@ public class User extends Model {
 	 * ユーザベクトルを取得する
 	 * @return
 	 */
-	public int[] getUserVector(int size) throws SQLException {
+	public double[] getUserVector(int size) throws SQLException {
 		return userService.getUserVector(this, size);
 	}
